улучшена навигация в приложении добавлен Observer для поиска

+++ MyApp/MyApp/Model/CoreData/Hourly+CoreDataProperties.swift
@@ -16,9 +16,9 @@ extension Hourly {
         return NSFetchRequest<Hourly>(entityName: "Hourly")
     }
 
-    @NSManaged public var temp: Int16
+    @NSManaged public var temp: Double
     @NSManaged public var name: String
-    @NSManaged public var fellsLike: Int16
+    @NSManaged public var fellsLike: Double
     @NSManaged public var date: Int64
     @NSManaged public var rain: Int16
     @NSManaged public var iconId: Int16
