+++ MyApp/MyApp/Model/CoreData/TopBar+CoreDataProperties.swift
@@ -17,11 +17,11 @@ extension TopBar {
     }
 
     @NSManaged public var date: Int64
-    @NSManaged public var feelsLike: Int16
+    @NSManaged public var feelsLike: Double
     @NSManaged public var desc: String?
     @NSManaged public var iconId: Int16
     @NSManaged public var pressure: Int16
-    @NSManaged public var temperature: Int16
+    @NSManaged public var temperature: Double
     @NSManaged public var weather: MainInfo?
 
 }
